task №1/requirements.txt
requests>=2.31.0
lxml>=5.1.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
